requirements.txt
streamlit
requests
newspaper3k
langdetect
pdfkit
lxml
lxml_html_clean
